Skipping new file event if file is already in .originals repo

+++ src/main/java/org/intellij/sdk/codesync/Utils.java
@@ -129,6 +129,8 @@ public class Utils {
         File f_originals = new File(destOriginals);
         File f_shadow = new File(destShadow);
 
+        if (f_originals.exists()) { return; }
+
         try {
             Files.copy(file.toPath(), f_originals.toPath());
         } catch (FileAlreadyExistsException e) {
